adding spec in correct location
